+++ src/MicroElements.Processing/TaskManager/Errors/Errors.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using MicroElements.Functional;
+
 namespace MicroElements.Processing.TaskManager
 {
     /// <summary>
@@ -8,32 +10,32 @@ namespace MicroElements.Processing.TaskManager
     /// </summary>
     public static class Errors
     {
-        public static IError<ErrorCode> SessionDoesNotExists(OperationId sessionId) => Error.CreateError(
+        public static Error<ErrorCode> SessionDoesNotExists(OperationId sessionId) => Error.CreateError(
             errorCode: ErrorCode.SessionDoesNotExists,
             messageTemplate: "Session does not exists. SessionId: {sessionId}.",
             args: sessionId);
 
-        public static IError<ErrorCode> SessionIsNotStarted(OperationId sessionId) => Error.CreateError(
+        public static Error<ErrorCode> SessionIsNotStarted(OperationId sessionId) => Error.CreateError(
             errorCode: ErrorCode.SessionIsNotStarted,
             messageTemplate: "Session is not started. SessionId: {sessionId}.",
             args: sessionId);
 
-        public static IError<ErrorCode> SessionIsAlreadyStarted(OperationId sessionId) => Error.CreateError(
+        public static Error<ErrorCode> SessionIsAlreadyStarted(OperationId sessionId) => Error.CreateError(
             errorCode: ErrorCode.SessionIsAlreadyStarted,
             messageTemplate: "Session is already started. SessionId: {sessionId}.",
             args: sessionId);
 
-        public static IError<ErrorCode> SessionUpdateIsProhibited(OperationId sessionId, string sessionStatus) => Error.CreateError(
+        public static Error<ErrorCode> SessionUpdateIsProhibited(OperationId sessionId, string sessionStatus) => Error.CreateError(
             errorCode: ErrorCode.SessionUpdateIsProhibited,
             messageTemplate: "Session update is prohibited. You can change session only in NotStarted status. SessionId: {sessionId}, SessionStatus: {sessionStatus}.",
             args: new object[] { sessionId, sessionStatus });
 
-        public static IError<ErrorCode> OperationDoesNotExists(OperationId operationId) => Error.CreateError(
+        public static Error<ErrorCode> OperationDoesNotExists(OperationId operationId) => Error.CreateError(
             errorCode: ErrorCode.OperationDoesNotExists,
             messageTemplate: "Operation does not exists. OperationId: {operationId}.",
             args: operationId);
 
-        public static IError<ErrorCode> OperationIdDoesNotMatch(OperationId providedOperationId, OperationId existingOperationId) => Error.CreateError(
+        public static Error<ErrorCode> OperationIdDoesNotMatch(OperationId providedOperationId, OperationId existingOperationId) => Error.CreateError(
             errorCode: ErrorCode.OperationIdDoesNotMatch,
             messageTemplate: "Provided OperationId does not match existing OperationId. ProvidedOperationId: {providedOperationId}, ExistingOperationId: {existingOperationId}.",
             args: new object[] { providedOperationId, existingOperationId });