+++ src/MicroElements.Processing/TaskManager/Errors/OperationManagerException.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Runtime.Serialization;
+using MicroElements.Functional;
 
 namespace MicroElements.Processing.TaskManager
 {
@@ -16,7 +17,7 @@ namespace MicroElements.Processing.TaskManager
         /// Initializes a new instance of the <see cref="OperationManagerException"/> class.
         /// </summary>
         /// <param name="error">Error associated with exception.</param>
-        public OperationManagerException(IError<ErrorCode> error)
+        public OperationManagerException(Error<ErrorCode> error)
             : base(error)
         {
         }